support query params as arrays in layer-composer
