+++ libs/layer-composer/src/generators/heatmap/util/fetch-breaks.ts
@@ -5,7 +5,7 @@ import { GlobalHeatmapAnimatedGeneratorConfig } from '../heatmap-animated'
 import { HeatmapAnimatedMode } from '../../types'
 import { isUrlAbsolute } from '../../../utils'
 import { Interval } from '../types'
-import { toURLArray } from '.'
+import { toURLArray } from '../../utils'
 
 export type Breaks = number[][]
 
@@ -169,8 +169,8 @@ export default function fetchBreaks(config: FetchBreaksParams): Promise<Breaks> 
       }
       const defaultDatasetKeys = Object.keys(BREAKS_FALLBACK) as DefaultDatasets[]
       const breaks = allDatasets.map((dataset) => {
-        const defaultDataset = defaultDatasetKeys.find((defaultDataset) =>
-          dataset?.includes(defaultDataset)
+        const defaultDataset = defaultDatasetKeys.find(
+          (defaultDataset) => dataset?.includes(defaultDataset)
         )
         return defaultDataset
           ? BREAKS_FALLBACK[defaultDataset]
